Renamed Mesh to MeshOGL. MeshOGL implements now Mesh.
