--- engine/src/main/java/pl/oblivion/engine/mesh/MeshData.java
@@ -1,16 +1,0 @@
-package pl.oblivion.engine.mesh;
-
-import java.util.List;
-
-public interface MeshData {
-
-  int getId();
-
-  int generateId();
-
-  List<VBO> getVBOS();
-
-  VBO getIndexVBO();
-
-  int getIndexCount();
-}
